--- test.c
@@ -1,5 +1,0 @@
-
-int main()
-{
-    return 0;
-}
